Fix Dockerfile builds to the wrong folder.
According to the Tomcat documentation the webapps folder needs to go
into .  is defined in the docker image that
is being used but is not referenced in the Dockerfile. Although the
Dockerfile builds without failure all request to the app fail with 404
because the sloan-ws.jar file is not added to the correct folder.

This change updates the Dockerfile to use /webapps as the
build target for sloan-ws.jar.

It also removes the unnecessary unpacking step of the jar file since
this gets automatically done by tomcat the first time the user attempts
to view the app.